refactor: extract shared audio capture module
- Create src/meeting/shared/audio-capture.ts with factory pattern
- Meet and Teams now use shared module with provider-specific config
- Reduces ~600 lines of duplication to ~30 lines per provider
- Combines best features: Meet's abort controller + Teams' periodic scanning
